hw4/requirements.txt
certifi==2021.5.30
cycler==0.10.0
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.21.4
pandas==1.1.5
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
#torch==1.9.0
torch==1.2.0
torchvision==0.4.0
#torchvision==0.10.0
tqdm==4.28.1
